+++ src/js/content/upgrades/base.js
@@ -1,6 +1,9 @@
 content.upgrades.base = {
   level: 0,
   levels: [],
+  canDowngrade: function () {
+    return this.level > 0
+  },
   canUpgrade: function () {
     const next = this.getNextLevel()
 
@@ -13,10 +16,17 @@ content.upgrades.base = {
   describe: () => '',
   describeNext: function () {
     if (!this.getNextLevel()) {
-      return 'Fully upgraded'
+      return
     }
 
     return this.describe(this.level + 1)
+  },
+  describePrevious: function () {
+    if (!this.getPreviousLevel()) {
+      return
+    }
+
+    return this.describe(this.level - 1)
   },
   getBonus: function () {
     const current = this.getLevel()
@@ -35,6 +45,9 @@ content.upgrades.base = {
   },
   getNextLevel: function () {
     return this.levels[this.level + 1]
+  },
+  getPreviousLevel: function () {
+    return this.levels[this.level - 1]
   },
   getProgress: function () {
     return this.level / (this.levels.length - 1)